Initializing Project
12/31/23

So far I have created a class WordExaminer that right now is extremely basic, and I don't see how complicated that in particular has to be. I'm also initializing a folder for manually-typed alterations to words tailored to help the AI dialect pronounce words more clearly, with less of a strong accent and less confusion in the specificied contexts. Lastly I've began working on a Txt object class, that will take an "opened()" txt file, and use that to organize the entered book text in ways that will both make referring to it in the primary script palatable, and in general making it simpler to present information as it needs to be divided for the Read Aloud feature anyways. (like pausing after the end of each paragraph, etc.)